command+core: Minimal initial implementation of -replace=... option
This only includes the internal mechanisms to make it work, and not any
of the necessary UI changes to "terraform plan" and "terraform apply" to
make their output make sense in this mode.

The force-replace options are ultimately handled inside the
NodeAbstractResourceInstance.plan method, at the same place we handle the
similar situation of the provider indicating that replacement is needed,
and so the rest of the changes here are just to propagate the settings
through all of the layers in order to reach that point.

Since the plan renderer has no awareness of this situation right now, it
renders force-replaced instances in the same way as once where the
provider indicates replacement is needed, annotated with "must be
replaced". In a later commit we'll add a special case to show these ones
as "will be replaced as requested", but otherwise this is a complete
implementation of the necessary functionality.