website: Reworking of the "terraform plan" docs, and related pages
It's been a long time since we gave this page an overhaul, and with our
ongoing efforts to make plan and apply incorporate all of the side-effects
that might need to be done against a configuration it seems like a good
time for some restructuring in that vein.

The starting idea here is to formally split the many "terraform plan"
options into a few different categories:
 - Planning modes
 - Planning options
 - Other options

The planning modes and options are the subset that are also accepted by
"terraform apply" when it's running in its default mode of generating a
plan and then prompting for interactive approval of it. This then allows
us to avoid duplicating all of that information on the "terraform apply"
page, and thus allows us to spend more words discussing each of them.

This also changes the docs for "terraform refresh" and "terraform taint"
to frame them as legacy commands preserved for backward compatibility,
pointing users instead to the relevant options on "terraform plan".

This commit describes some new options that were added in previous commits
but for which we plan some further improvements in subsequent commits.
The new "terraform plan" options described here are therefore not fully
functional as of this commit, but they have been minimally implemented
in the last two commits before this one.